pyproject.toml
##########################################
#                  RUFF                  #
##########################################
[tool.ruff]
exclude = [
    # Ruff standard
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    # MyFiles
    "__pycache__/*",
    "alembic/*",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "C",   # mccabe
    "N",   # pep8-naming
    "I",   # isort
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "Q",   # flake8-quotes
]
ignore = []

# Настройки flake8-quotes
[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

# # Настройки форматирования (замена black)
[tool.ruff.format]
quote-style = "single"
docstring-code-format = true
docstring-code-line-length = "dynamic"


##########################################
#                 Package                #
##########################################
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "universal-llm-service"
version = "1.0.17"
authors = [{ name = "Денис Шахбазян", email = "otbephncb@gmail.com" }]
description = "Библиотека для работы с языковыми моделями"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "aiohttp>=3.12.14",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-cerebras>=0.5.0",
    "langchain-deepseek>=0.1.3",
    "langchain-gigachat>=0.3.11",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "langchain-xai>=0.2.4",
    "tiktoken>=0.9.0",
    "tokenizers>=0.21.2",
]

[project.urls]
"Homepage" = "https://github.com/DenisShahbazyan/LLM_Service"
"Bug Tracker" = "https://github.com/DenisShahbazyan/LLM_Service/issues"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "build>=1.2.2.post1",
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.3",
    "pre-commit>=4.2.0",
    "python-dotenv>=1.1.1",
    "twine>=6.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["llm"]

[tool.hatch.build.targets.wheel.force-include]
"llm/tokenizer/deepseek/tokenizer.json" = "llm/tokenizer/deepseek/tokenizer.json"
